Fix de clases que usan a UserEntity.
Se actualizan los métodos utilizados en otras clases para coincidir con
los declarados de acuerdo a la tabla.

+++ src/main/java/com/lp/repository/UserRepository.java
@@ -2,12 +2,14 @@ package com.lp.repository;
 
 import com.lp.entity.UserEntity;
 
+import java.math.BigInteger;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import org.springframework.data.jpa.repository.Query;
 
 @Repository
-public interface UserRepository extends JpaRepository<UserEntity, Long>{
+public interface UserRepository extends JpaRepository<UserEntity, BigInteger>{
 
 	@Query("SELECT u FROM UserEntity u WHERE u.email = ?1")
 	public UserEntity findByEmail(String email);
